requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.4
pytest~=8.3.5
h11~=0.14.0
pip~=24.0
protobuf~=5.29.4
dlib~=19.24.6
grpcio~=1.71.0
grpcio-tools~=1.71.0
typing_extensions~=4.13.2
sniffio~=1.3.1
setuptools~=78.1.0
pluggy~=1.5.0
iniconfig~=2.1.0
fastapi~=0.115.12
starlette~=0.46.2
pydantic~=2.11.3
pydantic_core~=2.33.1
anyio~=4.9.0
click~=8.1.8
uvicorn~=0.34.2
annotated-types~=0.7.0
packaging~=25.0
face-recognition~=1.3.0
scikit-learn~=1.5.0